refactor: рефакториг кода
