+++ .eslintrc.cjs
@@ -1,5 +1,5 @@
 /* eslint-env node */
-require('@rushstack/eslint-patch/modern-module-resolution');
+require('@rushstack/eslint-patch/modern-module-resolution')
 
 module.exports = {
   root: true,
@@ -13,4 +13,4 @@ module.exports = {
   parserOptions: {
     ecmaVersion: 'latest',
   },
-};
+}
