--- src/components/Elements/Logo.vue
@@ -1,25 +1,0 @@
-<template>
-
-<a class="logo" href="#">
-   <img src="@/assets/images/icons/logo.svg" alt="логотип" width="121" height="30">
-</a>
-
-</template>
-
-<script>
-
-export default{
-
-}
-
-</script>
-
-<style>
-
-.logo {
-   width: 121px;
-   flex-shrink: 0;
-   display: inline-flex;
-}
-
-</style>
