+++ src/components/Elements/LogoElement.vue
@@ -1,0 +1,14 @@
+<template>
+  <router-link class="logo" to="/">
+    <img src="@/assets/images/icons/logo.svg" alt="логотип" width="121" height="30">
+  </router-link>
+
+</template>
+
+<style>
+.logo {
+  width: 121px;
+  flex-shrink: 0;
+  display: inline-flex;
+}
+</style>
