+++ src/router/index.ts
@@ -1,15 +1,21 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+import { createRouter, createWebHistory } from 'vue-router';
+import HomeView from '../views/HomeView.vue';
+import MapView from '../views/MapView.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: 'home',
-      component: HomeView
+      name: 'Home',
+      component: HomeView,
     },
-  ]
-})
+    {
+      path: '/map',
+      name: 'Map',
+      component: MapView,
+    },
+  ],
+});
 
-export default router
+export default router;
