+++ src/views/MapView.vue
@@ -1,0 +1,2 @@
+<template>
+</template>
